# docs/connectors.openapi.yaml
openapi: 3.1.0
info:
  title: Pulse Connectors API
  version: 1.0.0
  description: Organization-scoped API for provider-agnostic notification connectors.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /organizations/{orgId}/connectors/types:
    get:
      tags: [Connectors]
      summary: List registered connector provider types.
      parameters: [$ref: '#/components/parameters/OrgId']
      responses:
        '200': {$ref: '#/components/responses/ConnectorTypes'}
  /organizations/{orgId}/connectors:
    post:
      tags: [Connectors]
      summary: Create a connector.
      parameters: [$ref: '#/components/parameters/OrgId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateConnector'}
      responses:
        '201': {$ref: '#/components/responses/Connector'}
    get:
      tags: [Connectors]
      summary: List connectors.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - {name: type, in: query, schema: {$ref: '#/components/schemas/ConnectorType'}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ConnectorStatus'}}
        - {name: search, in: query, schema: {type: string}}
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': {$ref: '#/components/responses/ConnectorList'}
  /organizations/{orgId}/connectors/{id}:
    get:
      tags: [Connectors]
      summary: Get a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/Connector'}
    patch:
      tags: [Connectors]
      summary: Update a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateConnector'}
      responses:
        '200': {$ref: '#/components/responses/Connector'}
    delete:
      tags: [Connectors]
      summary: Soft-delete a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '204': {description: Deleted}
  /organizations/{orgId}/connectors/{id}/details:
    get:
      tags: [Connectors]
      summary: Get connector details.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/Connector'}
  /organizations/{orgId}/connectors/{id}/enable:
    post:
      tags: [Connectors]
      summary: Enable a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/Connector'}
  /organizations/{orgId}/connectors/{id}/disable:
    post:
      tags: [Connectors]
      summary: Disable a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/Connector'}
  /organizations/{orgId}/connectors/{id}/rotate-secret:
    post:
      tags: [Connector Operations]
      summary: Rotate connector credentials.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: {$ref: '#/components/schemas/ProviderConfig'}
      responses:
        '200': {$ref: '#/components/responses/Connector'}
  /organizations/{orgId}/connectors/{id}/test:
    post:
      tags: [Connector Operations]
      summary: Test provider connectivity.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/ConnectionTest'}
...
  /organizations/{orgId}/connectors/{id}/test-runs:
    get:
      tags: [Connector Operations]
      summary: List connector test runs.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId', $ref: '#/components/parameters/Limit', $ref: '#/components/parameters/Offset']
      responses:
        '200': {$ref: '#/components/responses/TestRunList'}
  /organizations/{orgId}/connectors/{id}/send:
    post:
      tags: [Connector Operations]
      summary: Send a test notification.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendNotification'}
      responses:
        '200': {$ref: '#/components/responses/DispatchSummary'}
        '502': {$ref: '#/components/responses/DispatchSummary'}
...
  /organizations/{orgId}/connectors/{id}/audit:
    get:
      tags: [Connector Audit]
      summary: List audit rows for a connector.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId', $ref: '#/components/parameters/Limit', $ref: '#/components/parameters/Offset']
      responses:
        '200': {$ref: '#/components/responses/AuditList'}
  /organizations/{orgId}/connectors/audit:
    get:
      tags: [Connector Audit]
      summary: List organization connector audit rows.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/Limit', $ref: '#/components/parameters/Offset']
      responses:
        '200': {$ref: '#/components/responses/AuditList'}
  /organizations/{orgId}/connectors/{id}/routes:
    post:
      tags: [Connector Routes]
      summary: Create a connector route.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConnectorRouteInput'}
      responses:
        '201': {$ref: '#/components/responses/ConnectorRoute'}
    get:
      tags: [Connector Routes]
      summary: List connector routes.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId', $ref: '#/components/parameters/Limit', $ref: '#/components/parameters/Offset']
      responses:
        '200': {$ref: '#/components/responses/ConnectorRouteList'}
  /organizations/{orgId}/connectors/{id}/routes/{routeId}:
    patch:
      tags: [Connector Routes]
      summary: Update a connector route.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId', $ref: '#/components/parameters/RouteId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConnectorRoutePatch'}
      responses:
        '200': {$ref: '#/components/responses/ConnectorRoute'}
    delete:
      tags: [Connector Routes]
      summary: Delete a connector route.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId', $ref: '#/components/parameters/RouteId']
      responses:
        '204': {description: Deleted}
  /organizations/{orgId}/connectors/{id}/oauth/start:
    post:
      tags: [Connector OAuth]
      summary: Start OAuth state and PKCE flow.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/OAuthStart'}
  /organizations/{orgId}/connectors/{id}/oauth/callback:
    post:
      tags: [Connector OAuth]
      summary: Validate and consume OAuth callback state.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OAuthCallback'}
      responses:
        '200':
          description: OAuth callback result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [connected, refreshQueued]
                        properties:
                          connected: {type: boolean}
                          refreshQueued: {type: boolean}
                          refreshJobId: {type: [string, 'null']}
  /organizations/{orgId}/connectors/{id}/oauth/refresh:
    post:
      tags: [Connector OAuth]
      summary: Request OAuth credential refresh.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200': {$ref: '#/components/responses/Queued'}
  /organizations/{orgId}/connectors/{id}/oauth/disconnect:
    post:
      tags: [Connector OAuth]
      summary: Revoke stored OAuth credentials.
      description: Archives the current OAuth credential version, replaces the live credential with an encrypted revoked marker, and marks the connector revoked.
      parameters: [$ref: '#/components/parameters/OrgId', $ref: '#/components/parameters/ConnectorId']
      responses:
        '200':
          description: Disconnect result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [disconnected]
                        properties:
                          disconnected: {type: boolean}
  /organizations/{orgId}/connectors/preview:
    post:
      tags: [Connector Operations]
      summary: Preview normalized notification payload.
      parameters: [$ref: '#/components/parameters/OrgId']
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendNotification'}
      responses:
        '200':
          description: Preview payload.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/PreviewNotification'}
  /organizations/{orgId}/connectors/validate-configuration:
    post:
      tags: [Connector Operations]
      summary: Validate provider configuration without saving.
      parameters: [$ref: '#/components/parameters/OrgId']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, config]
              properties:
                type: {$ref: '#/components/schemas/ConnectorType'}
                config: {$ref: '#/components/schemas/ProviderConfig'}
      responses:
        '200':
          description: Validation result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/ValidationResult'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId: {name: orgId, in: path, required: true, schema: {type: string, format: uuid}}
    ConnectorId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    RouteId: {name: routeId, in: path, required: true, schema: {type: string, format: uuid}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
  schemas:
    ConnectorType:
      type: string
      enum: [slack, discord, teams, pagerduty, webhook, email, sms]
    ConnectorStatus:
      type: string
      enum: [pending_setup, active, inactive, disabled, expired, revoked, degraded, error, rate_limited]
...
    ProviderConfig:
      type: object
      additionalProperties: true
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: {type: boolean, const: true}
...
    CreateConnector:
      type: object
      required: [name, type, config]
      properties:
        name: {type: string, minLength: 1}
        type: {$ref: '#/components/schemas/ConnectorType'}
        description: {type: string}
        config: {$ref: '#/components/schemas/ProviderConfig'}
        displayConfig: {type: object, additionalProperties: true}
        rateLimitRequests: {type: integer, minimum: 1}
        rateLimitWindowSeconds: {type: integer, minimum: 1}
        maxRetries: {type: integer, minimum: 0}
        failureThreshold: {type: integer, minimum: 1}
        metadata: {type: object, additionalProperties: true}
    UpdateConnector:
      type: object
      properties:
        name: {type: string, minLength: 1}
        description: {type: [string, 'null']}
        status: {$ref: '#/components/schemas/ConnectorStatus'}
        config: {$ref: '#/components/schemas/ProviderConfig'}
        displayConfig: {type: object, additionalProperties: true}
        rateLimitRequests: {type: integer, minimum: 1}
        rateLimitWindowSeconds: {type: integer, minimum: 1}
        maxRetries: {type: integer, minimum: 0}
        failureThreshold: {type: integer, minimum: 1}
        metadata: {type: object, additionalProperties: true}
    SendNotification:
      type: object
      properties:
        notificationType: {type: string, default: test}
        severity: {$ref: '#/components/schemas/Severity'}
        title: {type: string}
        body: {type: string}
        url: {type: string}
        fields:
          type: array
          items:
            type: object
            properties:
              label: {type: string}
              value: {type: string}
              short: {type: boolean}
        metadata: {type: object, additionalProperties: true}
...
    ConnectorRouteInput:
      type: object
      required: [eventType]
      properties:
        projectId: {type: [string, 'null'], format: uuid}
        environment:
          oneOf:
            - $ref: '#/components/schemas/ConnectorRouteEnvironment'
            - type: 'null'
        eventType: {type: string}
        severity: {$ref: '#/components/schemas/Severity'}
        enabled: {type: boolean}
    ConnectorRoutePatch:
      type: object
      properties:
        projectId: {type: [string, 'null'], format: uuid}
        environment:
          oneOf:
            - $ref: '#/components/schemas/ConnectorRouteEnvironment'
            - type: 'null'
        eventType: {type: string}
        severity: {$ref: '#/components/schemas/Severity'}
        enabled: {type: boolean}
    OAuthCallback:
      type: object
      required: [state]
      properties:
        state: {type: string}
        code: {type: string}
        error: {type: string}
        accessToken:
          type: string
          description: Provider access token. Stored encrypted and never returned.
        refreshToken:
          type: string
          description: Provider refresh token. Stored encrypted and never returned.
        tokenType:
          type: string
          default: Bearer
        scope:
          type: string
        expiresIn:
          type: integer
          minimum: 1
          maximum: 31536000
        expiresAt:
          type: string
          format: date-time
...
    ValidationResult:
      type: object
      required: [valid, errors]
      properties:
        valid: {type: boolean}
        errors: {type: array, items: {type: string}}
        normalized: {type: object, additionalProperties: true}
...
  responses:
    Connector:
      description: Connector response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {$ref: '#/components/schemas/Connector'}
    ConnectorList:
      description: Connector list response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Connector'}}
                  meta: {$ref: '#/components/schemas/ProviderConfig'}
    ConnectorTypes:
      description: Connector type catalog.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/ProviderConfig'}}
    ConnectionTest:
      description: Connection test result.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data:
                    type: object
                    required: [success, message, latencyMs]
                    properties:
                      success: {type: boolean}
                      message: {type: string}
                      latencyMs: {type: integer}
                      details: {type: object, additionalProperties: true}
...
    TestRunList:
      description: Test run list.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/ConnectorTestRun'}}
                  meta: {$ref: '#/components/schemas/ProviderConfig'}
    DispatchSummary:
      description: Dispatch summary.
      content:
        application/json:
          schema:
            type: object
            required: [success, data]
            properties:
              success: {type: boolean}
              data: {$ref: '#/components/schemas/DispatchSummary'}
...
    AuditList:
      description: Audit log list.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/AuditLog'}}
                  meta: {$ref: '#/components/schemas/ProviderConfig'}
    ConnectorRoute:
      description: Connector route response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {$ref: '#/components/schemas/ConnectorRoute'}
    ConnectorRouteList:
      description: Connector route list.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/ConnectorRoute'}}
                  meta: {$ref: '#/components/schemas/ProviderConfig'}
    OAuthStart:
      description: OAuth state response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data: {$ref: '#/components/schemas/OAuthStart'}
...